requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
apify-client>=1.8.0
gliner>=0.2.5
rapidfuzz>=3.6.0
pyyaml>=6.0
fastapi>=0.115.0
uvicorn>=0.30.0
